Fixes intermittent broken pipe error
The binary being executed didn't expect anything on stdin, so
sometimes its execution was finished before we even write something on
that stdin, leading to a write in a broken pipe.

Now the binary do a "read" command to force the use of stdin, as it is
expected to do so.

Fixes #12.

Signed-off-by: Giulian Gonçalves Vivan <giulian@ossystems.com.br>