install-if-different partial implementation
This change includes the major logic of the install-if-different
feature. What is still missing is the specific implementations of
comparations between the target and the sha256sum or pattern received
through json. FIXMEs were left in the code indicating this.

I think it is better to merge this partial implementation now because
this impacts directly on the architecture. The part missing won't
affect nothing else so we can develop it with less conflicts.

Signed-off-by: Giulian Gonçalves Vivan <giulian@ossystems.com.br>