implement validate and rollback procedures
They are procedures and not states, they are not reportable and
they run only once before the state machine starts.

When a "installed" state is achieved, it writes in the settings
the "UpgradeToInstallation"=? pointing to the new installation
index.

On the agent start, both validate and rollback states are ignored
when the "UpgradeToInstallation" is < 0 (-1 is the default value
when it's not set).

If "UpgradeToInstallation" is >= 0 AND is different from the current
active installation, a rollback procedure is launched. It just calls
the rollback callback and then go back to the state machine.

If "UpgradeToInstallation" is >= 0 AND is equal to the current active
installation, a validate procedure is launched. The validate procedure
calls the validate callback, if the callback succeeds, then it goes back
to the state machine. If the callback fails, it switches the active
partition and force a reboot.

Signed-off-by: Giulian Gonçalves Vivan <giulian@ossystems.com.br>