ensure that the same object can be downloaded twice
Since libarchive doesn't support seek, we need to be careful when
using the "libarchive.Reader" struct. We need te create a new Reader
struct every time we do an extraction so we ensure the internal
pointer of the archive is reset. On uhupkg files this doesn't impact
on performance since it uses "zip" format on which the files are
indexed.

The server backend implementation was already correct but a test was
missing to ensure that.

Signed-off-by: Giulian Gonçalves Vivan <giulian@ossystems.com.br>