firmware: tests: Allow easier creation of fake metadata
For tests which require the metadata to exist, it is important to have
an easy way to generate them and avoid duplicating a lot of code.

We did this providing a small API to generate the specific structure
so the normal loading procedure works as it was running on a real
system.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>