license: Relicense from GPL-2.0-only to MPL-2.0
The MPL is a simple copyleft license. The MPL's "file-level" copyleft
is designed to encourage contributors to share modifications they
make, while still allowing them to combine your code with code under
other licenses (open or proprietary) with minimal restrictions.

The MPL fills a useful space in the spectrum of free and open source
software licenses, sitting between the Apache license, which does not
require modifications to be shared, and the GNU family of licenses,
which requires modifications to be shared under a much broader set of
circumstances than the MPL, while remains compatible with a wide range
of licenses.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>