states: Handle state change callbacks in `State`
`State::handle_with_callback` verifies the return of the callback
script to decide if it goes to next state or to `Idle`.

The support has been added to following states:

- Download
- Install
- Reboot

This commit has been passed on the initial proposal[1] of Elias Gabriel
Amaral da Silva <tolkiendili@gmail.com>.

1. https://github.com/otavio/updatehub/pull/2

We reworked the proposed code to use a new trait and also added few
tests to ensures that the `TransitionCallback` trait is implemented for
all required states.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>