client: Reduce dependency of structures
Before this commit, the client API was storing references for
Settings, RuntimeSettings and Metadata. It made the API hard to use in
more flexible ways as this kept references for big structures while
using few parts of it.

This reworked the code so we pass the information as parameter and
avoid those huge struct references.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>