states: Avoid implementing default handlers for API
The probe and download_abort handlers should use a default
implementation so we reduce the code duplication.

When doing so, we ended fixing a problem of which states can handle
the respective requests.

The /update/download/abort should be accepted only inside the download
state. However, the /probe must be accepted on park and idle
states. Any other state should return as busy.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>