general: Update 'surf' crate to 2.0.0 release
The surf 2.0.0 release is out and this commit does the required code
changes to adapt to it. This should have no behavior change and all code
changes were done in internal code.

A side effect of the upgrade is the bump of the MSRV to 1.45.0. This
commit does the required changes on the GitHub Actions so it provides a
fully working commit.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>