updatehub: utils: definitions: Rework to reduce duplicated code
We have the `TargetTypeExt::get_device()` trait method which does all
the common work of figuring which device node to use, and the
`TargetTypeExt::valid()` method needs to ensure we are in good state to
use it.

The code has been reworked to rely in `get_device()` for the first part,
removing a lot of duplicated code in there.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>