updatehub: runtime_settings: backward compatibility with v1
When upgrading a device from UpdateHub v1 we need to load the context
from the disk, however, in case of an image validation error we need to
rollback to the previous image.

The rollback process, when coming from a v1 context, is different as we
need to _restore_ the old runtime settings content prior rebooting the
device.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>
Signed-off-by: Jonathas-Conceicao <jonathas.conceicao@ossystems.com.br>